+++ payment-service/src/main/java/onlinestore/paymentservice/dto/OrderDetailDto.java
@@ -4,13 +4,15 @@ import lombok.Builder;
 import lombok.Data;
 import onlinestore.paymentservice.model.entity.OrderDetailEntity;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class OrderDetailDto {
     private String productArticle;
-    private Double price;
+    private BigDecimal price;
     private Double quantity;
-    private Double amount;
+    private BigDecimal amount;
 
     public static OrderDetailDto fromOrderDetailEntity(OrderDetailEntity detailEntity) {
         return builder()
